+++ src/stock_trading/core/tests/mixins.py
@@ -379,3 +379,21 @@ class APITestCaseMixin:
         """
         response = self.client.patch(url, data)
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST, message)
+
+    def assertDisabledAction(
+            self,
+            url,
+            data={},
+            message="Method not allowed, action performed should disabled."
+    ):
+        """
+        Action is disabled.
+
+        :param url:
+        :param data:
+        :param message:
+        :return:
+        """
+        response = self.client.patch(url, data)
+        self.assertEqual(response.status_code, status.HTTP_405_METHOD_NOT_ALLOWED, message)
+